insights-engine_function-source/requirements.txt
google-cloud-bigquery
google-cloud-firestore
functions-framework

